Fixing Vulnerabilities

+++ src/Data/Services/TagService.cs
@@ -71,10 +71,12 @@ public class TagService(IDbContextFactory<ApplicationDbContext> dbContextFactory
     {
         using ApplicationDbContext context = await dbContextFactory.CreateDbContextAsync();
 
+
         Tag? foundTag = await context.Tags.FirstOrDefaultAsync(t => t.Id == tag.Id);
 
         if (foundTag is null) return;
 
+        if (await accessControlService.CheckUserAccessRightsRead(foundTag.IdProject) is false) return;
         context.Tags.Remove(foundTag);
         await context.SaveChangesAsync();
     }
